+++ src/client/slices/note.ts
@@ -90,7 +90,7 @@ const noteSlice = createSlice({
     removeAllNoteItalics: (state, { payload }: PayloadAction<NoteItem>) => {
       state.notes = state.notes.map((note) =>
         note.id === payload.id
-          ? { ...note, text: payload.text.replaceAll(/\*[\w*!\s]/ig, 'replace'), lastUpdated: payload.lastUpdated }
+          ? { ...note, text: payload.text.replaceAll(/\*+/g, (match, index)=> { return match.length === 1  ? '' : '**' }), lastUpdated: payload.lastUpdated }
           : note
       )
     },
